+++ dynamodb/src/test/scala/zio/dynamodb/codec/ItemDecoderSpec.scala
@@ -9,7 +9,7 @@ import java.time.Instant
 import scala.collection.immutable.ListMap
 
 object ItemDecoderSpec extends ZIOSpecDefault with CodecTestFixtures {
-  override def spec = suite("ItemDecoder Suite")(mainSuite)
+  override def spec = suite("ItemDecoder Suite")(mainSuite, noDiscriminatorSuite)
 
   val mainSuite = suite("Decoder Suite")(
     test("decodes generic record") {
@@ -313,6 +313,67 @@ object ItemDecoderSpec extends ZIOSpecDefault with CodecTestFixtures {
       val actual = DynamoDBQuery.fromItem[WithEnumWithoutDiscriminator](item)
 
       assert(actual)(isRight(equalTo(WithEnumWithoutDiscriminator(WithEnumWithoutDiscriminator.ONE))))
+    }
+  )
+
+  val noDiscriminatorSuite = suite("@noDisriminator Decoder Suite")(
+    test("decodes One case class with @fieldName") {
+      val item: Item = Item("sumType" -> Item("count" -> 42))
+
+      val actual = DynamoDBQuery.fromItem[WithNoDiscriminator](item)
+
+      assert(actual)(isRight(equalTo(WithNoDiscriminator(NoDiscriminatorEnum.One(i = 42)))))
+    },
+    test("decodes Two case class and ignores class level @caseName (which only applies to discriminators)") {
+      val item: Item = Item("sumType" -> Item("s" -> "X"))
+
+      val actual = DynamoDBQuery.fromItem[WithNoDiscriminator](item)
+
+      assert(actual)(isRight(equalTo(WithNoDiscriminator(NoDiscriminatorEnum.Two(s = "X")))))
+    },
+    test("decodes MinusOne case object") {
+      val item: Item = Item(Map("sumType" -> AttributeValue.String("MinusOne")))
+
+      val actual = DynamoDBQuery.fromItem[WithNoDiscriminator](item)
+
+      assert(actual)(isRight(equalTo(WithNoDiscriminator(NoDiscriminatorEnum.MinusOne))))
+    },
+    test("decodes Zero case object with @caseName") {
+      val item: Item = Item(Map("sumType" -> AttributeValue.String("0")))
+
+      val actual = DynamoDBQuery.fromItem[WithNoDiscriminator](item)
+
+      assert(actual)(isRight(equalTo(WithNoDiscriminator(NoDiscriminatorEnum.Zero))))
+    },
+    test("returns a Left of DecodeError when no decoder is not found") {
+      val item: Item = Item("sumType" -> Item("FIELD_NOT_IN_ANY_SUBTYPE" -> "X"))
+
+      val actual = DynamoDBQuery.fromItem[WithNoDiscriminator](item)
+
+      assert(actual)(
+        isLeft(
+          equalTo(
+            DynamoDBError.DecodingError(message =
+              "All sub type decoders failed for Map(Map(String(FIELD_NOT_IN_ANY_SUBTYPE) -> String(X)))"
+            )
+          )
+        )
+      )
+    },
+    test("returns a Left of DecodeError when a more than one decoder is found") {
+      val item: Item = Item("sumType" -> Item("i" -> 42))
+
+      val actual = DynamoDBQuery.fromItem[WithNoDiscriminatorError](item)
+
+      assert(actual)(
+        isLeft(
+          equalTo(
+            DynamoDBError.DecodingError(message =
+              "More than one sub type decoder succeeded for Map(Map(String(i) -> Number(42)))"
+            )
+          )
+        )
+      )
     }
   )
